Clean up Generic Tuple Sketches to be more consistent.
