+++ src/main/java/org/apache/datasketches/tuple/adouble/DoubleSummaryFactory.java
@@ -23,18 +23,12 @@ import org.apache.datasketches.tuple.SummaryFactory;
 
 /**
  * Factory for DoubleSummary.
+ *
+ * @author Lee Rhodes
  */
 public final class DoubleSummaryFactory implements SummaryFactory<DoubleSummary> {
 
   private final DoubleSummary.Mode summaryMode_;
-
-  /**
-   * Creates an instance of DoubleSummaryFactory with default mode
-   */
-  @Deprecated
-  public DoubleSummaryFactory() {
-    summaryMode_ = DoubleSummary.Mode.Sum;
-  }
 
   /**
    * Creates an instance of DoubleSummaryFactory with a given mode
